isolate _create_study function.
